requirements.txt
openai==1.58.1
psycopg[binary]==3.1.18
sqlalchemy==2.0.36
python-dotenv==1.0.1
streamlit==1.41.1
pandas==2.2.3
